Move assignTerminal to module
